2048 game made using C++
Our project 2048 is a 16 cells – 1 grid game. Each cell contains one tile of a number. That number is a multiple of 2k, where k = 1,2,3,4… So the numbers on the tiles will be 2, 4, 8, 16, 32, 64, 128, 256, 512, 1024, 2048, 4096, 8192. When you start it for first time (as a matter of fact, for every new game), two random tiles of numbers 2 or will occupy the cells. You just have to swipe up / down / right / left. When two tiles of the same number meet, they are merged and form one tile equal to the two times number of previous tiles. Example – Tile of numbers 2 and 2 will make the tile of number 4. This game is very simple to play. This game can be played by even a 4th standard student.